requirements.txt
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
selenium~=4.16.0
tinydb==4.8.2
SQLAlchemy~=2.0.36